refactored update pfp
