+++ src/components/auth/SignupForm.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import SocialLoginButtons from "../SocialLoginButtons";
+import SocialLoginButtons from "./SocialLoginButtons";
 import Spinner from "../spinners/Spinner";
 import { Link } from "react-router-dom";
 import { AuthServices } from "../../api/auth";
