+++ src/components/auth/SocialLoginButtons.jsx
@@ -1,6 +1,9 @@
 export default function SocialLoginButtons() {
     const continueWithGoogle = () => {
-        window.open("/api/v1/auth/google", "_self");
+        window.open(
+            `${import.meta.env.VITE_SERVER_URL}/api/v1/auth/google`,
+            "_self"
+        );
     };
     return (
         <div className="flex-1 max-w-md">
